music player history on destroy

+++ cogs/music_player.py
@@ -110,7 +110,7 @@ def build_queue_embed(player):
     return embed
 
 
-def build_history_embed(player):
+def build_history_embed(player, title: str):
     text = ""
     if not player.history:
         text = "*Empty*"
@@ -124,7 +124,7 @@ def build_history_embed(player):
             if len(text) > 3000:
                 break
     embed = Embed(description=text)
-    embed.set_author(name="History:", icon_url="https://cdn.discordapp.com/emojis/695126168680005662.webp")
+    embed.set_author(name=title, icon_url="https://cdn.discordapp.com/emojis/695126168680005662.webp")
     return embed
 
 
@@ -193,6 +193,7 @@ class ExtPlayer(wavelink.Player):
         self.message = None
         self.controls = None
         self.ctx = None
+        self.hash = None
 
 
 # class PlayerView(View):
@@ -384,7 +385,7 @@ class PlayerView(View):
             self.player.queue.clear()
             await self.player.stop()
         elif custom_id == 'history':
-            await interaction.response.send_message(embed=build_history_embed(self.player), ephemeral=True)
+            await interaction.response.send_message(embed=build_history_embed(self.player, "History:"), ephemeral=True)
             return True
         elif custom_id == 'queue_list':
             value = interaction.data['values'][0]
@@ -436,19 +437,24 @@ class MusicPlayerCog(commands.Cog, name="Music player"):
 
     @commands.Cog.listener()
     async def on_wavelink_track_start(self, player: ExtPlayer, track: Track):
+        player.hash = random.random()
         if player.message is None:
             await self.player_message(player)
 
     async def soft_destroy(self, player):
+        h = player.hash
         await asyncio.sleep(10)
+        if h != player.hash:
+            return
         if not player.is_playing():
             try:
                 await player.disconnect()
                 await player.message.delete()
+                await player.ctx.send(embed=build_history_embed(player, "Songs played during the session:"))
+                if player.guild.id in self.players and self.players[player.guild.id] == player:
+                    self.players.pop(player.guild.id)
             except:
                 pass
-            if player.guild.id in self.players and self.players[player.guild.id] == player:
-                self.players.pop(player.guild.id)
 
     @commands.Cog.listener()
     async def on_wavelink_track_end(self, player: ExtPlayer, track: Track, reason):
@@ -700,7 +706,8 @@ class MusicPlayerCog(commands.Cog, name="Music player"):
     async def history(self, ctx):
         server_id = ctx.guild.id
         if server_id in self.players:
-            await ctx.response.send_message(embed=build_history_embed(self.players[server_id]),
+            await ctx.response.send_message(embed=build_history_embed(self.players[server_id],
+                                                                      "History:"),
                                             delete_after=60.0, ephemeral=True)
         else:
             await ctx.response.send_message(Embed(title="Player not initialized", color=Colour.red()),
